pyproject.toml
[project]
name = "dspy-exploration"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.84.0",
    "black>=26.3.1",
    "dspy>=3.1.3",
    "google-api-python-client>=2.194.0",
    "google-auth-httplib2>=0.3.1",
    "google-auth-oauthlib>=1.3.1",
    "notebook>=7.5.5",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
    "tavily-python>=0.7.23",
    "yfinance>=1.3.0",
]
